requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
websocket-client==1.6.4
python-dotenv==1.0.0
talib==0.4.28
scikit-learn==1.3.2
joblib==1.3.2
pyyaml==6.0.1
asyncio==3.4.3
aiohttp==3.9.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
fastapi==0.104.1
uvicorn==0.24.0
